Extracted TabbedPaneIcon class from JChessTabbedPane class
