+++ src/main/java/com/ovgu/ccd/gui/Moves.java
@@ -28,7 +28,6 @@ import com.ovgu.ccd.pieces.Square;
 
 import javax.swing.*;
 import javax.swing.table.AbstractTableModel;
-import javax.swing.table.DefaultTableModel;
 import java.awt.*;
 import java.util.ArrayList;
 import java.util.Stack;
